third_party/WebKit/LayoutTests/fast/dynamic/hover-after-affected-by-change.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  .affected:hover { color: green }
  #hoveredContents { display: contents }
</style>
<div id="hovered">Hover me - should become green</div>
<div id="hoveredContents">
  <div id="hovered2">Hover me - should become green</div>
</div>
<script>
  function mouseMoveTo(element) {
    return new Promise((resolve, reject) => {
      chrome.gpuBenchmarking.pointerActionSequence([{source: "mouse",
          actions: [{ name: "pointerMove",
              x: element.offsetLeft + 1,
              y: element.offsetTop + 1 }]}],
          resolve);
    });
  }

  test(() => {
    hovered.offsetTop;
    hovered.className = "affected";
    hoveredContents.className = "affected";
  }, "Change selector matching causes different AffectedByHover flags.");

  test(() => {
    assert_exists(window, "chrome", "This test requires window.chrome.");
    assert_exists(window.chrome, "gpuBenchmarking",
        "This test requires chrome.gpuBenchmarking.");
  }, "Preconditions.");

  promise_test(async () => {
    assert_equals(getComputedStyle(hovered).color, "rgb(0, 0, 0)");
    await mouseMoveTo(hovered);
    assert_equals(getComputedStyle(hovered).color, "rgb(0, 128, 0)");
  }, "Check that :hover rule is applied.");

  promise_test(async () => {
    assert_equals(getComputedStyle(hoveredContents).color, "rgb(0, 0, 0)");
    await mouseMoveTo(hovered2);
    assert_equals(getComputedStyle(hoveredContents).color, "rgb(0, 128, 0)");
  }, "Check that :hover rule is applied to display:contents element.");
</script>
